Datebank wieder zum laufen gebracht

+++ sep/src/main/java/com/gruppe_f/sep/entities/liga/LigaRepository.java
@@ -2,7 +2,7 @@ package com.gruppe_f.sep.entities.liga;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LigaRepository extends JpaRepository<Liga, Integer> {
+public interface LigaRepository extends JpaRepository<Liga, Long> {
 
     Liga findLigaById(long id);
